create header file, move demo to examples
