--- Makefile
@@ -1,3 +1,0 @@
-all:
-	g++ -g3 idstring.cpp -o idstring
-	gdb -ex run ./idstring
