+++ examples/basic/Makefile
@@ -1,0 +1,4 @@
+all: main
+
+main: main.cpp ../../idstring.cpp
+	g++ -o main -I../../ main.cpp ../../idstring.cpp
